refactor: complete OptionRowLHN decomposition with ArchivedOptionRowLHN variant
